requirements.txt
playwright>=1.44.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.31.0
python-dateutil>=2.9.0
